% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{article_svm}
\alias{article_svm}
\title{SVM article}
\usage{
article_svm(...)
}
\arguments{
\item{...}{}
}
\value{
An R Markdown output format.
}
\description{
SVM article
}
\details{
This was adapted from \url{https://github.com/svmiller/svm-r-markdown-templates}.
}
