% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{copy_rmd}
\alias{copy_rmd}
\title{Copy the existing .Rmd file from the rosr package to the destination directory.}
\usage{
copy_rmd(package, template, sub_project, rmd_dir)
}
\arguments{
\item{package}{Packge name of the source template.}

\item{template}{Template name of the source template.}

\item{sub_project}{Sub-project name}

\item{rmd_dir}{Destination directory}
}
\value{
copied file and the path of the new .Rmd file
}
\description{
Copy the existing .Rmd file from the rosr package to the destination directory.
}
