% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_bib}
\alias{create_bib}
\title{Create a demo .bib file}
\usage{
create_bib(to = getwd())
}
\arguments{
\item{to}{the destination directory}
}
\value{
a demo .bib file
}
\description{
Create a demo .bib file
}
\examples{
\dontrun{
create_bib()
}
}
