% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_data}
\alias{create_data}
\title{Create a demo data file}
\usage{
create_data(to = getwd())
}
\arguments{
\item{to}{the destination directory}
}
\value{
a demo data file
}
\description{
Create a demo data file
}
\examples{
create_data(tempdir())
}
