% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_dir}
\alias{create_dir}
\title{Create an empty directory}
\usage{
create_dir(to = getwd())
}
\arguments{
\item{to}{the destination directory}
}
\value{
an empty directory
}
\description{
Create an empty directory
}
\examples{
create_dir()
}
