% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_eq}
\alias{create_eq}
\title{Create a demo equation sheet}
\usage{
create_eq(to = getwd(), if_render = TRUE)
}
\arguments{
\item{to}{the destination directory}

\item{if_render}{Logical. Whether render the templates.}
}
\value{
a demo equation sheet
}
\description{
Create a demo equation sheet
}
\examples{
\dontrun{
create_eq()
}
}
