% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_mindmap}
\alias{create_mindmap}
\title{Create a demo mind map}
\usage{
create_mindmap(to = getwd(), if_save = FALSE)
}
\arguments{
\item{to}{the destination directory}

\item{if_save}{Logical. Whether save the mindmap.}
}
\value{
a demo mind map
}
\description{
Create a demo mind map
}
\examples{
create_mindmap(tempdir())
}
