% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_r}
\alias{create_r}
\title{Create a demo .R script}
\usage{
create_r(to = getwd())
}
\arguments{
\item{to}{the destination directory}
}
\value{
a demo .R script
}
\description{
Create a demo .R script
}
\examples{
create_r(tempdir())
}
