% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_rmd}
\alias{create_rmd}
\title{Create a demo markdown project}
\usage{
create_rmd(to = "manuscript", template = "copernicus_article", if_render = TRUE, 
    package = "rticles")
}
\arguments{
\item{to}{the desnitation directory}

\item{template}{the name of the r markdown template. It is the folder name in 'package/rmarkdown/templates/'.}

\item{if_render}{whether render it.}

\item{package}{the package name where the template comes from.}
}
\value{
copied file.
}
\description{
Create a demo markdown project
}
\examples{
\donttest{
create_rmd(tempdir(), if_render = FALSE)
}
}
