% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_rosr}
\alias{create_rosr}
\title{Create a new project or a new sub-project}
\usage{
create_rosr(project = "test", if_render = TRUE, dest_dir = NA, overwrite = FALSE, 
    sub_project = sub_projects())
}
\arguments{
\item{project}{Project name.}

\item{if_render}{Logical. Whether render the templates.}

\item{dest_dir}{Character. The destination directory of the new project.}

\item{overwrite}{Logical. Whether overwrite existing files.}

\item{sub_project}{Character. The type of the sub-project to create.}
}
\value{
a project folder
}
\description{
Create a new project or a new sub-project
}
\examples{
create_rosr(dest_dir = tempdir(), sub_project = "Rproj")
}
