% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_rpkg}
\alias{create_rpkg}
\title{Create a demo R package}
\usage{
create_rpkg(to = getwd())
}
\arguments{
\item{to}{the destination directory}
}
\value{
a demo R package
}
\description{
Create a demo R package
}
\examples{
create_rpkg(tempdir())
}
