% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{create_ui}
\alias{create_ui}
\title{A shiny app for creating a rosr project}
\usage{
create_ui()
}
\value{
a shinyapp which can be displayed in a web browser.
}
\description{
A shiny app for creating a rosr project
}
\examples{
\dontrun{
create_ui()
}
}
