% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_website}
\alias{create_website}
\title{Create a demo blogdown website}
\usage{
create_website(to = getwd(), theme = "yihui/hugo-lithium", oldwd = getwd())
}
\arguments{
\item{to}{the desnitation directory}

\item{theme}{the name of the r blogdown theme.}

\item{oldwd}{set the working directory aftering rendering the book}
}
\value{
copied file.
}
\description{
Create a demo blogdown website
}
\examples{
create_website(tempdir())
}
