% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{cv_svm}
\alias{cv_svm}
\title{SVM cv}
\usage{
cv_svm(...)
}
\arguments{
\item{...}{}
}
\value{
An R Markdown output format.
}
\description{
SVM cv
}
\details{
This was adapted from \url{https://github.com/svmiller/svm-r-markdown-templates}.
}
