% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{eq}
\alias{eq}
\title{Insert an equation.}
\usage{
eq(eqs = NULL, label = NULL, number = NULL, style = c("numbered", "display", "inline", 
    "none"))
}
\arguments{
\item{label}{Character. The label of the equation.}

\item{number}{integer. The number of the equation.}

\item{eq_df}{Character. The path to the equation text file, or a data frame of the equations.}
}
\value{
A string of the equation.
}
\description{
Insert an equation.
}
\examples{
eq()
}
