% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq_word}
\alias{eq_word}
\title{Convert a MathML equation for Microsoft Word}
\usage{
eq_word(connection = "clipboard")
}
\arguments{
\item{connection}{a connection object or a character string.}
}
\value{
a Word equation
}
\description{
Convert a MathML equation for Microsoft Word
}
\details{
Copy the MathML codes of an equation into your clipboard, and run eq_word(), then paste from  your clipboard to MS Word.
}
\examples{
# Copy the MathML codes of an equation into your clipboard.
eq_word(NA)
# Now paste to MS Word.
}
