% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{fig}
\alias{fig}
\title{Insert an image in the 'image/' directory.}
\usage{
fig(filename)
}
\arguments{
\item{filename}{The file name of the image.}
}
\value{
Insert an image.
}
\description{
Insert an image in the 'image/' directory.
}
