% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_sub_project}
\alias{get_sub_project}
\title{Display available sub projects}
\usage{
get_sub_project()
}
\value{
character of the project names
}
\description{
Display available sub projects
}
\examples{
get_sub_project()
}
