% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_yaml}
\alias{get_yaml}
\title{Get the yaml header of a string vecter from a markdown file}
\usage{
get_yaml(txt)
}
\arguments{
\item{txt}{a string vector vecter from a markdown file}
}
\value{
a string vector of the yaml header
}
\description{
Get the yaml header of a string vecter from a markdown file
}
