% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{install_packages}
\alias{install_packages}
\title{Install dependencies}
\usage{
install_packages()
}
\value{
install packages
}
\description{
Install dependencies
}
\examples{
\dontrun{
install_packages()
}
}
