% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{manuscript_svm}
\alias{manuscript_svm}
\title{SVM manuscript}
\usage{
manuscript_svm(...)
}
\arguments{
\item{...}{}
}
\value{
An R Markdown output format.
}
\description{
SVM manuscript
}
\details{
This was adapted from \url{https://github.com/svmiller/svm-r-markdown-templates}.
}
