% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read.eq}
\alias{read.eq}
\title{Read equations from a file}
\usage{
read.eq(eqs, skip)
}
\arguments{
\item{eqs}{character. The path of the equation file.}

\item{skip}{integer. the number of lines of the data file to skip before beginning to read data.}
}
\value{
a data frame with equations
}
\description{
Read equations from a file
}
\examples{
eq_file <- paste0(system.file(package = "rosr"), "/skeleton/equation/rosr-eq.Rmd")
eqs <- read.eq(eq_file)
}
