% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{read_eq}
\alias{read_eq}
\title{Read equations from a file}
\usage{
read_eq(eqs, skip = 6)
}
\arguments{
\item{eqs}{character. The path of the equation file.}

\item{skip}{integer. the number of lines of the data file to skip before beginning to read data.}
}
\value{
a data frame with equations
}
\description{
Read equations from a file
}
\examples{
\dontrun{
eq_file <- paste0(system.file(package = "rosr"), "/skeleton/equation/rosr-eq.Rmd")
eqs <- read_eq(eq_file)
}
}
