% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{rm_space}
\alias{rm_space}
\title{remove the spaces before and after a string.}
\usage{
rm_space(string)
}
\arguments{
\item{string}{a string.}
}
\value{
a new string.
}
\description{
remove the spaces before and after a string.
}
