% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{rmd_template}
\alias{rmd_template}
\title{rosr Rmd template}
\usage{
rmd_template(..., template_name, keep_tex = TRUE, citation_package = "natbib", 
    md_extensions = c("-autolink_bare_uris", "-auto_identifiers"))
}
\arguments{
\item{...}{Additional arguments to \code{rmarkdown::pdf_document()}.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}
}
\value{
An R Markdown output format.

An R Markdown output format.
}
\description{
rosr Rmd template
}
