\name{rosr-imports}
\alias{clipr_available}
\alias{write_clip}
\alias{a}
\alias{br}
\alias{code}
\alias{div}
\alias{em}
\alias{h1}
\alias{h2}
\alias{h3}
\alias{h4}
\alias{h5}
\alias{h6}
\alias{hr}
\alias{HTML}
\alias{htmlTemplate}
\alias{img}
\alias{includeCSS}
\alias{includeHTML}
\alias{includeMarkdown}
\alias{includeScript}
\alias{includeText}
\alias{is.singleton}
\alias{knit_print.html}
\alias{knit_print.shiny.tag}
\alias{knit_print.shiny.tag.list}
\alias{p}
\alias{pre}
\alias{singleton}
\alias{span}
\alias{strong}
\alias{suppressDependencies}
\alias{tag}
\alias{tagAppendAttributes}
\alias{tagAppendChild}
\alias{tagAppendChildren}
\alias{tagList}
\alias{tagSetChildren}
\alias{validateCssUnit}
\alias{withTags}
\alias{..stacktraceoff..}
\alias{..stacktraceon..}
\alias{absolutePanel}
\alias{actionButton}
\alias{actionLink}
\alias{addResourcePath}
\alias{animationOptions}
\alias{appendTab}
\alias{as.shiny.appobj}
\alias{basicPage}
\alias{bookmarkButton}
\alias{bootstrapLib}
\alias{bootstrapPage}
\alias{browserViewer}
\alias{brushedPoints}
\alias{brushOpts}
\alias{callModule}
\alias{captureStackTraces}
\alias{checkboxGroupInput}
\alias{checkboxInput}
\alias{clickOpts}
\alias{column}
\alias{conditionalPanel}
\alias{conditionStackTrace}
\alias{conditionStackTrace<-}
\alias{createRenderFunction}
\alias{createWebDependency}
\alias{dataTableOutput}
\alias{dateInput}
\alias{dateRangeInput}
\alias{dblclickOpts}
\alias{debounce}
\alias{dialogViewer}
\alias{diskCache}
\alias{downloadButton}
\alias{downloadHandler}
\alias{downloadLink}
\alias{enableBookmarking}
\alias{eventReactive}
\alias{exportTestValues}
\alias{exprToFunction}
\alias{extractStackTrace}
\alias{fileInput}
\alias{fillCol}
\alias{fillPage}
\alias{fillRow}
\alias{fixedPage}
\alias{fixedPanel}
\alias{fixedRow}
\alias{flowLayout}
\alias{fluidPage}
\alias{fluidRow}
\alias{formatStackTrace}
\alias{freezeReactiveVal}
\alias{freezeReactiveValue}
\alias{getCurrentOutputInfo}
\alias{getDefaultReactiveDomain}
\alias{getQueryString}
\alias{getShinyOption}
\alias{getUrlHash}
\alias{headerPanel}
\alias{helpText}
\alias{hideTab}
\alias{hoverOpts}
\alias{htmlOutput}
\alias{icon}
\alias{imageOutput}
\alias{incProgress}
\alias{inputPanel}
\alias{insertTab}
\alias{insertUI}
\alias{installExprFunction}
\alias{invalidateLater}
\alias{is.key_missing}
\alias{is.reactive}
\alias{is.reactivevalues}
\alias{is.shiny.appobj}
\alias{isolate}
\alias{isRunning}
\alias{isTruthy}
\alias{key_missing}
\alias{knit_print.reactive}
\alias{knit_print.shiny.appobj}
\alias{knit_print.shiny.render.function}
\alias{mainPanel}
\alias{makeReactiveBinding}
\alias{markRenderFunction}
\alias{maskReactiveContext}
\alias{memoryCache}
\alias{modalButton}
\alias{modalDialog}
\alias{navbarMenu}
\alias{navbarPage}
\alias{navlistPanel}
\alias{nearPoints}
\alias{need}
\alias{NS}
\alias{ns.sep}
\alias{numericInput}
\alias{observe}
\alias{observeEvent}
\alias{onBookmark}
\alias{onBookmarked}
\alias{onFlush}
\alias{onFlushed}
\alias{onReactiveDomainEnded}
\alias{onRestore}
\alias{onRestored}
\alias{onSessionEnded}
\alias{onStop}
\alias{outputOptions}
\alias{pageWithSidebar}
\alias{paneViewer}
\alias{parseQueryString}
\alias{passwordInput}
\alias{plotOutput}
\alias{plotPNG}
\alias{prependTab}
\alias{printError}
\alias{printStackTrace}
\alias{Progress}
\alias{radioButtons}
\alias{reactive}
\alias{reactiveFileReader}
\alias{reactivePlot}
\alias{reactivePoll}
\alias{reactivePrint}
\alias{reactiveTable}
\alias{reactiveText}
\alias{reactiveTimer}
\alias{reactiveUI}
\alias{reactiveVal}
\alias{reactiveValues}
\alias{reactiveValuesToList}
\alias{registerInputHandler}
\alias{removeInputHandler}
\alias{removeModal}
\alias{removeNotification}
\alias{removeTab}
\alias{removeUI}
\alias{renderCachedPlot}
\alias{renderDataTable}
\alias{renderImage}
\alias{renderPlot}
\alias{renderPrint}
\alias{renderTable}
\alias{renderText}
\alias{renderUI}
\alias{repeatable}
\alias{req}
\alias{restoreInput}
\alias{runApp}
\alias{runExample}
\alias{runGadget}
\alias{runGist}
\alias{runGitHub}
\alias{runUrl}
\alias{safeError}
\alias{selectInput}
\alias{selectizeInput}
\alias{serverInfo}
\alias{setBookmarkExclude}
\alias{setProgress}
\alias{setSerializer}
\alias{shinyApp}
\alias{shinyAppDir}
\alias{shinyAppFile}
\alias{shinyOptions}
\alias{shinyServer}
\alias{shinyUI}
\alias{showBookmarkUrlModal}
\alias{showModal}
\alias{showNotification}
\alias{showReactLog}
\alias{showTab}
\alias{sidebarLayout}
\alias{sidebarPanel}
\alias{sizeGrowthRatio}
\alias{sliderInput}
\alias{snapshotExclude}
\alias{snapshotPreprocessInput}
\alias{snapshotPreprocessOutput}
\alias{splitLayout}
\alias{stopApp}
\alias{submitButton}
\alias{tableOutput}
\alias{tabPanel}
\alias{tabsetPanel}
\alias{tags}
\alias{textAreaInput}
\alias{textInput}
\alias{textOutput}
\alias{throttle}
\alias{titlePanel}
\alias{uiOutput}
\alias{updateActionButton}
\alias{updateCheckboxGroupInput}
\alias{updateCheckboxInput}
\alias{updateDateInput}
\alias{updateDateRangeInput}
\alias{updateNavbarPage}
\alias{updateNavlistPanel}
\alias{updateNumericInput}
\alias{updateQueryString}
\alias{updateRadioButtons}
\alias{updateSelectInput}
\alias{updateSelectizeInput}
\alias{updateSliderInput}
\alias{updateTabsetPanel}
\alias{updateTextAreaInput}
\alias{updateTextInput}
\alias{updateVarSelectInput}
\alias{updateVarSelectizeInput}
\alias{urlModal}
\alias{validate}
\alias{varSelectInput}
\alias{varSelectizeInput}
\alias{verbatimTextOutput}
\alias{verticalLayout}
\alias{wellPanel}
\alias{withLogErrors}
\alias{withMathJax}
\alias{withProgress}
\alias{withReactiveDomain}
\alias{install.packages}
\alias{read.table}
\alias{unzip}
\alias{write.csv}
\docType{import}
\title{Objects imported from other packages}
\description{
These objects are imported from other packages. Follow the links to their documentation.
\describe{
  \item{clipr}{\code{\link[clipr:clipr_available]{clipr_available}}, \code{\link[clipr:write_clip]{write_clip}}}
  \item{htmltools}{\code{\link[htmltools:a]{a}}, \code{\link[htmltools:br]{br}}, \code{\link[htmltools:code]{code}}, \code{\link[htmltools:div]{div}}, \code{\link[htmltools:em]{em}}, \code{\link[htmltools:h1]{h1}}, \code{\link[htmltools:h2]{h2}}, \code{\link[htmltools:h3]{h3}}, \code{\link[htmltools:h4]{h4}}, \code{\link[htmltools:h5]{h5}}, \code{\link[htmltools:h6]{h6}}, \code{\link[htmltools:hr]{hr}}, \code{\link[htmltools:HTML]{HTML}}, \code{\link[htmltools:htmlTemplate]{htmlTemplate}}, \code{\link[htmltools:img]{img}}, \code{\link[htmltools:includeCSS]{includeCSS}}, \code{\link[htmltools:includeHTML]{includeHTML}}, \code{\link[htmltools:includeMarkdown]{includeMarkdown}}, \code{\link[htmltools:includeScript]{includeScript}}, \code{\link[htmltools:includeText]{includeText}}, \code{\link[htmltools:is.singleton]{is.singleton}}, \code{\link[htmltools:knit_print.html]{knit_print.html}}, \code{\link[htmltools:knit_print.shiny.tag]{knit_print.shiny.tag}}, \code{\link[htmltools:knit_print.shiny.tag.list]{knit_print.shiny.tag.list}}, \code{\link[htmltools:p]{p}}, \code{\link[htmltools:pre]{pre}}, \code{\link[htmltools:singleton]{singleton}}, \code{\link[htmltools:span]{span}}, \code{\link[htmltools:strong]{strong}}, \code{\link[htmltools:suppressDependencies]{suppressDependencies}}, \code{\link[htmltools:tag]{tag}}, \code{\link[htmltools:tagAppendAttributes]{tagAppendAttributes}}, \code{\link[htmltools:tagAppendChild]{tagAppendChild}}, \code{\link[htmltools:tagAppendChildren]{tagAppendChildren}}, \code{\link[htmltools:tagList]{tagList}}, \code{\link[htmltools:tagSetChildren]{tagSetChildren}}, \code{\link[htmltools:validateCssUnit]{validateCssUnit}}, \code{\link[htmltools:withTags]{withTags}}}
  \item{shiny}{\code{\link[shiny:..stacktraceoff..]{..stacktraceoff..}}, \code{\link[shiny:..stacktraceon..]{..stacktraceon..}}, \code{\link[shiny:absolutePanel]{absolutePanel}}, \code{\link[shiny:actionButton]{actionButton}}, \code{\link[shiny:actionLink]{actionLink}}, \code{\link[shiny:addResourcePath]{addResourcePath}}, \code{\link[shiny:animationOptions]{animationOptions}}, \code{\link[shiny:appendTab]{appendTab}}, \code{\link[shiny:as.shiny.appobj]{as.shiny.appobj}}, \code{\link[shiny:basicPage]{basicPage}}, \code{\link[shiny:bookmarkButton]{bookmarkButton}}, \code{\link[shiny:bootstrapLib]{bootstrapLib}}, \code{\link[shiny:bootstrapPage]{bootstrapPage}}, \code{\link[shiny:browserViewer]{browserViewer}}, \code{\link[shiny:brushedPoints]{brushedPoints}}, \code{\link[shiny:brushOpts]{brushOpts}}, \code{\link[shiny:callModule]{callModule}}, \code{\link[shiny:captureStackTraces]{captureStackTraces}}, \code{\link[shiny:checkboxGroupInput]{checkboxGroupInput}}, \code{\link[shiny:checkboxInput]{checkboxInput}}, \code{\link[shiny:clickOpts]{clickOpts}}, \code{\link[shiny:column]{column}}, \code{\link[shiny:conditionalPanel]{conditionalPanel}}, \code{\link[shiny:conditionStackTrace]{conditionStackTrace}}, \code{\link[shiny:conditionStackTrace<-]{conditionStackTrace<-}}, \code{\link[shiny:createRenderFunction]{createRenderFunction}}, \code{\link[shiny:createWebDependency]{createWebDependency}}, \code{\link[shiny:dataTableOutput]{dataTableOutput}}, \code{\link[shiny:dateInput]{dateInput}}, \code{\link[shiny:dateRangeInput]{dateRangeInput}}, \code{\link[shiny:dblclickOpts]{dblclickOpts}}, \code{\link[shiny:debounce]{debounce}}, \code{\link[shiny:dialogViewer]{dialogViewer}}, \code{\link[shiny:diskCache]{diskCache}}, \code{\link[shiny:downloadButton]{downloadButton}}, \code{\link[shiny:downloadHandler]{downloadHandler}}, \code{\link[shiny:downloadLink]{downloadLink}}, \code{\link[shiny:enableBookmarking]{enableBookmarking}}, \code{\link[shiny:eventReactive]{eventReactive}}, \code{\link[shiny:exportTestValues]{exportTestValues}}, \code{\link[shiny:exprToFunction]{exprToFunction}}, \code{\link[shiny:extractStackTrace]{extractStackTrace}}, \code{\link[shiny:fileInput]{fileInput}}, \code{\link[shiny:fillCol]{fillCol}}, \code{\link[shiny:fillPage]{fillPage}}, \code{\link[shiny:fillRow]{fillRow}}, \code{\link[shiny:fixedPage]{fixedPage}}, \code{\link[shiny:fixedPanel]{fixedPanel}}, \code{\link[shiny:fixedRow]{fixedRow}}, \code{\link[shiny:flowLayout]{flowLayout}}, \code{\link[shiny:fluidPage]{fluidPage}}, \code{\link[shiny:fluidRow]{fluidRow}}, \code{\link[shiny:formatStackTrace]{formatStackTrace}}, \code{\link[shiny:freezeReactiveVal]{freezeReactiveVal}}, \code{\link[shiny:freezeReactiveValue]{freezeReactiveValue}}, \code{\link[shiny:getCurrentOutputInfo]{getCurrentOutputInfo}}, \code{\link[shiny:getDefaultReactiveDomain]{getDefaultReactiveDomain}}, \code{\link[shiny:getQueryString]{getQueryString}}, \code{\link[shiny:getShinyOption]{getShinyOption}}, \code{\link[shiny:getUrlHash]{getUrlHash}}, \code{\link[shiny:headerPanel]{headerPanel}}, \code{\link[shiny:helpText]{helpText}}, \code{\link[shiny:hideTab]{hideTab}}, \code{\link[shiny:hoverOpts]{hoverOpts}}, \code{\link[shiny:htmlOutput]{htmlOutput}}, \code{\link[shiny:icon]{icon}}, \code{\link[shiny:imageOutput]{imageOutput}}, \code{\link[shiny:incProgress]{incProgress}}, \code{\link[shiny:inputPanel]{inputPanel}}, \code{\link[shiny:insertTab]{insertTab}}, \code{\link[shiny:insertUI]{insertUI}}, \code{\link[shiny:installExprFunction]{installExprFunction}}, \code{\link[shiny:invalidateLater]{invalidateLater}}, \code{\link[shiny:is.key_missing]{is.key_missing}}, \code{\link[shiny:is.reactive]{is.reactive}}, \code{\link[shiny:is.reactivevalues]{is.reactivevalues}}, \code{\link[shiny:is.shiny.appobj]{is.shiny.appobj}}, \code{\link[shiny:isolate]{isolate}}, \code{\link[shiny:isRunning]{isRunning}}, \code{\link[shiny:isTruthy]{isTruthy}}, \code{\link[shiny:key_missing]{key_missing}}, \code{\link[shiny:knit_print.reactive]{knit_print.reactive}}, \code{\link[shiny:knit_print.shiny.appobj]{knit_print.shiny.appobj}}, \code{\link[shiny:knit_print.shiny.render.function]{knit_print.shiny.render.function}}, \code{\link[shiny:mainPanel]{mainPanel}}, \code{\link[shiny:makeReactiveBinding]{makeReactiveBinding}}, \code{\link[shiny:markRenderFunction]{markRenderFunction}}, \code{\link[shiny:maskReactiveContext]{maskReactiveContext}}, \code{\link[shiny:memoryCache]{memoryCache}}, \code{\link[shiny:modalButton]{modalButton}}, \code{\link[shiny:modalDialog]{modalDialog}}, \code{\link[shiny:navbarMenu]{navbarMenu}}, \code{\link[shiny:navbarPage]{navbarPage}}, \code{\link[shiny:navlistPanel]{navlistPanel}}, \code{\link[shiny:nearPoints]{nearPoints}}, \code{\link[shiny:need]{need}}, \code{\link[shiny:NS]{NS}}, \code{\link[shiny:ns.sep]{ns.sep}}, \code{\link[shiny:numericInput]{numericInput}}, \code{\link[shiny:observe]{observe}}, \code{\link[shiny:observeEvent]{observeEvent}}, \code{\link[shiny:onBookmark]{onBookmark}}, \code{\link[shiny:onBookmarked]{onBookmarked}}, \code{\link[shiny:onFlush]{onFlush}}, \code{\link[shiny:onFlushed]{onFlushed}}, \code{\link[shiny:onReactiveDomainEnded]{onReactiveDomainEnded}}, \code{\link[shiny:onRestore]{onRestore}}, \code{\link[shiny:onRestored]{onRestored}}, \code{\link[shiny:onSessionEnded]{onSessionEnded}}, \code{\link[shiny:onStop]{onStop}}, \code{\link[shiny:outputOptions]{outputOptions}}, \code{\link[shiny:pageWithSidebar]{pageWithSidebar}}, \code{\link[shiny:paneViewer]{paneViewer}}, \code{\link[shiny:parseQueryString]{parseQueryString}}, \code{\link[shiny:passwordInput]{passwordInput}}, \code{\link[shiny:plotOutput]{plotOutput}}, \code{\link[shiny:plotPNG]{plotPNG}}, \code{\link[shiny:prependTab]{prependTab}}, \code{\link[shiny:printError]{printError}}, \code{\link[shiny:printStackTrace]{printStackTrace}}, \code{\link[shiny:Progress]{Progress}}, \code{\link[shiny:radioButtons]{radioButtons}}, \code{\link[shiny:reactive]{reactive}}, \code{\link[shiny:reactiveFileReader]{reactiveFileReader}}, \code{\link[shiny:reactivePlot]{reactivePlot}}, \code{\link[shiny:reactivePoll]{reactivePoll}}, \code{\link[shiny:reactivePrint]{reactivePrint}}, \code{\link[shiny:reactiveTable]{reactiveTable}}, \code{\link[shiny:reactiveText]{reactiveText}}, \code{\link[shiny:reactiveTimer]{reactiveTimer}}, \code{\link[shiny:reactiveUI]{reactiveUI}}, \code{\link[shiny:reactiveVal]{reactiveVal}}, \code{\link[shiny:reactiveValues]{reactiveValues}}, \code{\link[shiny:reactiveValuesToList]{reactiveValuesToList}}, \code{\link[shiny:registerInputHandler]{registerInputHandler}}, \code{\link[shiny:removeInputHandler]{removeInputHandler}}, \code{\link[shiny:removeModal]{removeModal}}, \code{\link[shiny:removeNotification]{removeNotification}}, \code{\link[shiny:removeTab]{removeTab}}, \code{\link[shiny:removeUI]{removeUI}}, \code{\link[shiny:renderCachedPlot]{renderCachedPlot}}, \code{\link[shiny:renderDataTable]{renderDataTable}}, \code{\link[shiny:renderImage]{renderImage}}, \code{\link[shiny:renderPlot]{renderPlot}}, \code{\link[shiny:renderPrint]{renderPrint}}, \code{\link[shiny:renderTable]{renderTable}}, \code{\link[shiny:renderText]{renderText}}, \code{\link[shiny:renderUI]{renderUI}}, \code{\link[shiny:repeatable]{repeatable}}, \code{\link[shiny:req]{req}}, \code{\link[shiny:restoreInput]{restoreInput}}, \code{\link[shiny:runApp]{runApp}}, \code{\link[shiny:runExample]{runExample}}, \code{\link[shiny:runGadget]{runGadget}}, \code{\link[shiny:runGist]{runGist}}, \code{\link[shiny:runGitHub]{runGitHub}}, \code{\link[shiny:runUrl]{runUrl}}, \code{\link[shiny:safeError]{safeError}}, \code{\link[shiny:selectInput]{selectInput}}, \code{\link[shiny:selectizeInput]{selectizeInput}}, \code{\link[shiny:serverInfo]{serverInfo}}, \code{\link[shiny:setBookmarkExclude]{setBookmarkExclude}}, \code{\link[shiny:setProgress]{setProgress}}, \code{\link[shiny:setSerializer]{setSerializer}}, \code{\link[shiny:shinyApp]{shinyApp}}, \code{\link[shiny:shinyAppDir]{shinyAppDir}}, \code{\link[shiny:shinyAppFile]{shinyAppFile}}, \code{\link[shiny:shinyOptions]{shinyOptions}}, \code{\link[shiny:shinyServer]{shinyServer}}, \code{\link[shiny:shinyUI]{shinyUI}}, \code{\link[shiny:showBookmarkUrlModal]{showBookmarkUrlModal}}, \code{\link[shiny:showModal]{showModal}}, \code{\link[shiny:showNotification]{showNotification}}, \code{\link[shiny:showReactLog]{showReactLog}}, \code{\link[shiny:showTab]{showTab}}, \code{\link[shiny:sidebarLayout]{sidebarLayout}}, \code{\link[shiny:sidebarPanel]{sidebarPanel}}, \code{\link[shiny:sizeGrowthRatio]{sizeGrowthRatio}}, \code{\link[shiny:sliderInput]{sliderInput}}, \code{\link[shiny:snapshotExclude]{snapshotExclude}}, \code{\link[shiny:snapshotPreprocessInput]{snapshotPreprocessInput}}, \code{\link[shiny:snapshotPreprocessOutput]{snapshotPreprocessOutput}}, \code{\link[shiny:splitLayout]{splitLayout}}, \code{\link[shiny:stopApp]{stopApp}}, \code{\link[shiny:submitButton]{submitButton}}, \code{\link[shiny:tableOutput]{tableOutput}}, \code{\link[shiny:tabPanel]{tabPanel}}, \code{\link[shiny:tabsetPanel]{tabsetPanel}}, \code{\link[shiny:tags]{tags}}, \code{\link[shiny:textAreaInput]{textAreaInput}}, \code{\link[shiny:textInput]{textInput}}, \code{\link[shiny:textOutput]{textOutput}}, \code{\link[shiny:throttle]{throttle}}, \code{\link[shiny:titlePanel]{titlePanel}}, \code{\link[shiny:uiOutput]{uiOutput}}, \code{\link[shiny:updateActionButton]{updateActionButton}}, \code{\link[shiny:updateCheckboxGroupInput]{updateCheckboxGroupInput}}, \code{\link[shiny:updateCheckboxInput]{updateCheckboxInput}}, \code{\link[shiny:updateDateInput]{updateDateInput}}, \code{\link[shiny:updateDateRangeInput]{updateDateRangeInput}}, \code{\link[shiny:updateNavbarPage]{updateNavbarPage}}, \code{\link[shiny:updateNavlistPanel]{updateNavlistPanel}}, \code{\link[shiny:updateNumericInput]{updateNumericInput}}, \code{\link[shiny:updateQueryString]{updateQueryString}}, \code{\link[shiny:updateRadioButtons]{updateRadioButtons}}, \code{\link[shiny:updateSelectInput]{updateSelectInput}}, \code{\link[shiny:updateSelectizeInput]{updateSelectizeInput}}, \code{\link[shiny:updateSliderInput]{updateSliderInput}}, \code{\link[shiny:updateTabsetPanel]{updateTabsetPanel}}, \code{\link[shiny:updateTextAreaInput]{updateTextAreaInput}}, \code{\link[shiny:updateTextInput]{updateTextInput}}, \code{\link[shiny:updateVarSelectInput]{updateVarSelectInput}}, \code{\link[shiny:updateVarSelectizeInput]{updateVarSelectizeInput}}, \code{\link[shiny:urlModal]{urlModal}}, \code{\link[shiny:validate]{validate}}, \code{\link[shiny:varSelectInput]{varSelectInput}}, \code{\link[shiny:varSelectizeInput]{varSelectizeInput}}, \code{\link[shiny:verbatimTextOutput]{verbatimTextOutput}}, \code{\link[shiny:verticalLayout]{verticalLayout}}, \code{\link[shiny:wellPanel]{wellPanel}}, \code{\link[shiny:withLogErrors]{withLogErrors}}, \code{\link[shiny:withMathJax]{withMathJax}}, \code{\link[shiny:withProgress]{withProgress}}, \code{\link[shiny:withReactiveDomain]{withReactiveDomain}}}
  \item{utils}{\code{\link[utils:install.packages]{install.packages}}, \code{\link[utils:read.table]{read.table}}, \code{\link[utils:unzip]{unzip}}, \code{\link[utils:write.csv]{write.csv}}}
}}
