% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{rosr_server}
\alias{rosr_server}
\title{Server for the Shiny app create_ui}
\usage{
rosr_server(input, output, session)
}
\arguments{
\item{input}{The input of the server.}

\item{output}{The output of the server.}

\item{session}{The session of the server.}
}
\description{
Server for the Shiny app create_ui
}
