% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{statement_svm}
\alias{statement_svm}
\title{SVM statement}
\usage{
statement_svm(...)
}
\arguments{
\item{...}{}
}
\value{
An R Markdown output format.
}
\description{
SVM statement
}
\details{
This was adapted from \url{https://github.com/svmiller/svm-r-markdown-templates}.
}
