% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{sub_projects}
\alias{sub_projects}
\title{Display available sub projects}
\usage{
sub_projects()
}
\value{
character of the project names
}
\description{
Display available sub projects
}
\examples{
sub_projects()
}
