% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{syllabus_svm}
\alias{syllabus_svm}
\title{SVM syllabus}
\usage{
syllabus_svm(...)
}
\arguments{
\item{...}{}
}
\value{
An R Markdown output format.
}
\description{
SVM syllabus
}
\details{
This was adapted from \url{https://github.com/svmiller/svm-r-markdown-templates}.
}
