% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{template_ls}
\alias{template_ls}
\title{Display available templates in a list}
\usage{
template_ls()
}
\value{
a list of templates
}
\description{
Display available templates in a list
}
\examples{
\dontrun{
template_ls()
}
}
