% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{template_of_sub_project}
\alias{template_of_sub_project}
\title{Get the template name of the given sub-project name}
\usage{
template_of_sub_project(sub_project)
}
\arguments{
\item{sub_project}{the name of the sub-project}
}
\value{
the name of the template
}
\description{
Get the template name of the given sub-project name
}
