% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{templates}
\alias{templates}
\title{Display available templates in a data frame}
\usage{
templates()
}
\value{
a data frame of templates
}
\description{
Display available templates in a data frame
}
\examples{
templates()
}
