% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create_rproj}
\alias{create_rproj}
\title{Create a .Rproj file}
\usage{
create_rproj(project)
}
\arguments{
\item{project}{the name of the .Rproj file}
}
\value{
a .Rproj.file
}
\description{
Create a .Rproj file
}
\examples{
\dontrun{
create_rproj()
}
}
