% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/begin.R
\name{install_packages}
\alias{install_packages}
\title{Install dependencies}
\usage{
install_packages(server = c("cran", "github"))
}
\arguments{
\item{server}{character. The server from where to install the dependencies.}
}
\value{
install packages
}
\description{
Install dependencies
}
\examples{
install_packages(NA)
}
