% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd.R
\name{rmd_template}
\alias{rmd_template}
\alias{statement_svm}
\alias{article_svm}
\alias{cv_svm}
\alias{syllabus_svm}
\alias{manuscript_svm}
\alias{letter_moderncv}
\alias{beamer_zh}
\title{rosr Rmd template}
\usage{
rmd_template(..., template_name, keep_tex = TRUE, citation_package = "natbib", 
    latex_engine = "xelatex", md_extensions = c("-autolink_bare_uris", 
        "-auto_identifiers"))

statement_svm(...)

article_svm(...)

cv_svm(...)

syllabus_svm(...)

manuscript_svm(...)

letter_moderncv(...)

beamer_zh(...)
}
\arguments{
\item{..., }{keep_tex, citation_package, md_extensions
Arguments passed to \code{rmarkdown::pdf_document()}.}

\item{template_name}{the template name.}

\item{keep_tex}{Keep the intermediate tex file used in the conversion to PDF}

\item{citation_package}{The LaTeX package to process citations, \code{natbib}
or \code{biblatex}. Use \code{none} if neither package is to be used.}

\item{latex_engine}{LaTeX engine for producing PDF output. Options are
"pdflatex", "lualatex", and "xelatex".}

\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link{rmarkdown_format}} for
additional details.}
}
\value{
An R Markdown output format.
}
\description{
rosr Rmd template
}
\details{
statement_svm, article_svm, cv_svm, syllabus_svm, manuscript_svm are adapted from \url{https://github.com/svmiller/svm-r-markdown-templates}.
}
